requirements.txt
git+https://github.com/huggingface/transformers
accelerate >= 0.12.0
deepspeed
datasets >= 1.8.0
sentencepiece != 0.1.92
scipy
scikit-learn
protobuf
torch >= 1.3
evaluate
python-dotenv
wandb